🏗 Create gulp integration (#22673)
