+++ build-system/pr-check/remote-tests.js
@@ -47,8 +47,8 @@ async function main() {
     timedExecOrDie('gulp update-packages');
 
     await startSauceConnect(FILENAME);
-    timedExecOrDie('gulp unit --nobuild --saucelabs_lite');
-    timedExecOrDie('gulp test --integration --nobuild --compiled --saucelabs');
+    timedExecOrDie('gulp unit --nobuild --saucelabs');
+    timedExecOrDie('gulp integration --nobuild --compiled --saucelabs');
 
     stopSauceConnect(FILENAME);
   } else {
@@ -76,7 +76,7 @@ async function main() {
     await startSauceConnect(FILENAME);
 
     if (buildTargets.has('RUNTIME') || buildTargets.has('UNIT_TEST')) {
-      timedExecOrDie('gulp unit --nobuild --saucelabs_lite');
+      timedExecOrDie('gulp unit --nobuild --saucelabs');
     }
 
     if (
@@ -84,9 +84,7 @@ async function main() {
       buildTargets.has('FLAG_CONFIG') ||
       buildTargets.has('INTEGRATION_TEST')
     ) {
-      timedExecOrDie(
-        'gulp test --integration --nobuild --compiled --saucelabs'
-      );
+      timedExecOrDie('gulp integration --nobuild --compiled --saucelabs');
     }
     stopSauceConnect(FILENAME);
   }
