+++ build-system/tasks/runtime-test/index.js
@@ -27,12 +27,6 @@ const opn = require('opn');
 const path = require('path');
 const webserver = require('gulp-webserver');
 const {
-  getAdTypes,
-  isLargeRefactor,
-  refreshKarmaWdCache,
-  unitTestsToRun,
-} = require('./helpers');
-const {
   reportTestErrored,
   reportTestFinished,
   reportTestRunComplete,
@@ -44,6 +38,8 @@ const {clean} = require('../clean');
 const {createCtrlcHandler, exitCtrlcHandler} = require('../../ctrlcHandler');
 const {css} = require('../css');
 const {dist} = require('../dist');
+const {getAdTypes, refreshKarmaWdCache} = require('./helpers');
+const {isLargeRefactor, unitTestsToRun} = require('./helpers-unit');
 const {isTravisBuild} = require('../../travis');
 
 const {green, yellow, cyan, red} = colors;
