+++ gulpfile.js
@@ -47,6 +47,7 @@ const {dist} = require('./build-system/tasks/dist');
 const {e2e} = require('./build-system/tasks/e2e');
 const {firebase} = require('./build-system/tasks/firebase');
 const {getZindex} = require('./build-system/tasks/get-zindex');
+const {integration} = require('./build-system/tasks/integration');
 const {lint} = require('./build-system/tasks/lint');
 const {makeExtension} = require('./build-system/tasks/extension-generator');
 const {nailgunStart, nailgunStop} = require('./build-system/tasks/nailgun');
@@ -88,6 +89,7 @@ gulp.task('dist', dist);
 gulp.task('e2e', e2e);
 gulp.task('firebase', firebase);
 gulp.task('get-zindex', getZindex);
+gulp.task('integration', integration);
 gulp.task('json-syntax', jsonSyntax);
 gulp.task('lint', lint);
 gulp.task('make-extension', makeExtension);
